+++ app/helpers/application_helper.rb
@@ -5,6 +5,7 @@ module ApplicationHelper
   include ::FinancialAssistance::VerificationHelper
   include HtmlScrubberUtil
   include DropdownHelper
+  include RelationshipDropdownHelper
 
   def add_external_links_enabled?
     EnrollRegistry.feature_enabled?(:add_external_links)
@@ -612,15 +613,6 @@ module ApplicationHelper
                     end
 
     options_for_select(relationships.map{|r| [display_relationship(r.to_s), r.to_s] }, selected: dependent.try(:relationship))
-  end
-
-  def display_relationship(relationship)
-    relationship_mapping = {
-      'domestic_partners_child' => "Domestic partner's child",
-      'parents_domestic_partner' => "Parent's domestic partner"
-    }
-
-    relationship_mapping[relationship].try(:humanize) || relationship.try(:humanize)
   end
 
   def enrollment_progress_bar(plan_year, p_min, options = {:minimum => true})
